+++ CHANGELOG.md
@@ -13,7 +13,8 @@
 | KarteNotificationServiceExtension | リッチプッシュ通知機能を提供します。 | 1.3.0 |
 | KarteDebugger | KARTE for Appのデバッグ機能を提供します。 | 1.1.0 |
 
-# Releases - xxxx.xx.xx
+# Releases - 2025.12.09
+## Version 2.31.0
 
 ### Core 2.34.0
 **💊FIXED**